requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.7
langchain-google-genai==2.0.8
langgraph==0.2.53
pydantic==2.9.2
python-dotenv==1.0.1
